scenario 2 - di
