+++ app/src/main/java/com/overlord/defenses/Weapon.java
@@ -1,0 +1,5 @@
+package com.overlord.defenses;
+
+public interface Weapon {
+    String fire();
+}
