+++ app/src/main/java/com/overlord/villagers/Builder.java
@@ -1,0 +1,7 @@
+package com.overlord.villagers;
+
+public class Builder extends Villager {
+    public String buildSomething() {
+        return "Building a house";
+    }
+}
