+++ app/src/main/java/com/overlord/villagers/Hunter.java
@@ -1,0 +1,7 @@
+package com.overlord.villagers;
+
+public class Hunter extends Villager {
+    public String gatherFood() {
+        return "Hunting some deers";
+    }
+}
