+++ app/src/main/java/com/overlord/villagers/Lumberjack.java
@@ -1,0 +1,7 @@
+package com.overlord.villagers;
+
+public class Lumberjack extends Villager {
+    public String cutWood() {
+        return "Cutting some wood";
+    }
+}
