+++ app/src/test/java/com/overlord/defenses/GuardTowerTest.java
@@ -8,15 +8,14 @@ class GuardTowerTest {
 
     @Test
     public void build_your_defenses() {
-        // given
-        Ballista ballista = new Ballista();
-
         // when
-        // TODO consctruct your tower with a ballista
+        // TODO construct your tower with a ballista (the ballista is the default weapon in the constructor)
         GuardTower tower = new GuardTower();
         // TODO upgrade your tower with a cannon
+        Cannon cannon = new Cannon();
+        tower.setWeapon(cannon);
         // TODO fire some deadly projectile
-        String actualProjectile = "";
+        String actualProjectile = tower.fire();
 
         // then
         // TODO check that a cannonball has been fired
